requirements.txt
flask==3.0.3
pandas==2.2.2
python-dotenv==1.0.1
twilio==9.3.1
cohere==5.5.8
